Made a change to the setup of where the filter section sits in the components.
